Move snapping utils
